aiagents4pharma/talk2biomodels/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
 
[project]
name = "talk2biomodels"

description = "A submodule of AIAgents4Pharma for interacting with systems biology models "

readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "copasi_basico==0.78",
    "hydra-core==1.3.2",
    "openai==1.59.4",
    "pandas==2.2.3",
    "langchain==0.3.7",
    "langchain-community==0.3.5",
    "langchain-core==0.3.15",
    "langchain-experimental==0.3.3",
    "langchain-openai==0.2.5",
    "langgraph==0.2.62",
    "matplotlib==3.9.2",
    "plotly==5.24.1",
    "pydantic==2.9.2",
    "pylint==3.3.1",
    "pytest==8.3.3",
    "streamlit==1.39.0"
]
dynamic = ["version"]
 
# Package discovery for the submodule
[tool.setuptools]
packages = [
    "talk2biomodels",
    "talk2biomodels.models",
    "talk2biomodels.tools",
    "talk2biomodels.agents",
    "talk2biomodels.states"
]